fix bugs

+++ routes/web.php
@@ -26,19 +26,6 @@ use Illuminate\Support\Facades\Route;
 
 require __DIR__.'/auth.php';
 
-/*
-|--------------------------------------------------------------------------
-| Welcome Route
-|--------------------------------------------------------------------------
-*/
-//Route::get('/', function () {
-//    return Inertia::render('Welcome', [
-//        'canLogin' => Route::has('login'),
-//        'canRegister' => Route::has('register'),
-//        'laravelVersion' => Application::VERSION,
-//        'phpVersion' => PHP_VERSION,
-//    ]);
-//});
 
 /*
 |--------------------------------------------------------------------------
@@ -61,52 +48,49 @@ Route::get('/summary', [SummaryController::class, 'index'])
 | User Routes
 |--------------------------------------------------------------------------
 */
-Route::prefix('users')->middleware(['auth', 'admin'])->name('users.')->group(function () {
-    Route::get('/', [UserController::class, 'index'])->name('index');
-    Route::get('/create', [UserController::class, 'create'])->name('create');
-    Route::post('/', [UserController::class, 'store'])->name('store');
-    Route::get('/{user}', [UserController::class, 'show'])->name('show');
-    Route::get('/{user}/edit', [UserController::class, 'edit'])->name('edit');
-    Route::delete('/{user}', [UserController::class, 'destroy'])->name('destroy');
-});
+    Route::get('/users', [UserController::class, 'index'])->middleware(['auth', 'admin'])->name('users.index');
+    Route::get('/users/create', [UserController::class, 'create'])->middleware(['auth', 'admin'])->name('users.create');
+    Route::post('/users', [UserController::class, 'store'])->middleware(['auth', 'admin'])->name('users.store');
+    Route::get('/users/{user}', [UserController::class, 'show'])->middleware(['auth', 'admin'])->name('users.show');
+    Route::get('/users/{user}/edit', [UserController::class, 'edit'])->middleware(['auth', 'admin'])->name('users.edit');
+    Route::put('/users/{user}', [UserController::class, 'update'])->name('users.update');
+    Route::delete('/users/{user}', [UserController::class, 'destroy'])->middleware(['auth', 'admin'])->name('users.destroy');
+
 
 /*
 |--------------------------------------------------------------------------
 | Topic Routes
 |--------------------------------------------------------------------------
 */
-Route::prefix('topics')->middleware(['auth', 'admin'])->name('topics.')->group(function () {
-    Route::get('/', [TopicController::class, 'index'])->name('index');
-    Route::get('/create', [TopicController::class, 'create'])->name('create');
-    Route::post('/', [TopicController::class, 'store'])->name('store');
-    Route::get('/{topic}', [TopicController::class, 'show'])->name('show');
-    Route::get('/{topic}/edit', [TopicController::class, 'edit'])->name('edit');
-    Route::put('/{topic}', [TopicController::class, 'update'])->name('update');
-    Route::delete('/{topic}', [TopicController::class, 'destroy'])->name('destroy');
-});
+    Route::get('/topics', [TopicController::class, 'index'])->middleware(['auth', 'admin'])->name('topics.index');
+    Route::get('/topics/create', [TopicController::class, 'create'])->middleware(['auth', 'admin'])->name('topics.create');
+    Route::post('/topics', [TopicController::class, 'store'])->middleware(['auth', 'admin'])->name('topics.store');
+    Route::get('/topics/{topic}', [TopicController::class, 'show'])->middleware(['auth', 'admin'])->name('topics.show');
+    Route::get('/topics/{topic}/edit', [TopicController::class, 'edit'])->middleware(['auth', 'admin'])->name('topics.edit');
+    Route::put('/topics/{topic}', [TopicController::class, 'update'])->middleware(['auth', 'admin'])->name('topics.update');
+    Route::delete('/topics/{topic}', [TopicController::class, 'destroy'])->middleware(['auth', 'admin'])->name('topics.destroy');
+
 
 /*
 |--------------------------------------------------------------------------
 | Discussion Routes
 |--------------------------------------------------------------------------
 */
-Route::prefix('discussions')->middleware(['auth', 'admin'])->name('discussions.')->group(function () {
-    Route::get('/', [DiscussionController::class, 'index'])->name('index');
-    Route::get('/create', [DiscussionController::class, 'create'])->name('create');
-    Route::get('/{discussion:slug}/edit', [DiscussionController::class, 'edit'])->name('edit');
-    Route::put('/{discussion}', [DiscussionController::class, 'update'])->name('update');
+    Route::get('/discussions', [DiscussionController::class, 'index'])->middleware(['auth', 'admin'])->name('discussions.index');
+    Route::get('/discussions/create', [DiscussionController::class, 'create'])->middleware(['auth', 'admin'])->name('discussions.create');
+    Route::get('/discussions/{discussion:slug}/edit', [DiscussionController::class, 'edit'])->middleware(['auth', 'admin'])->name('discussions.edit');
+    Route::put('/discussions/{discussion}', [DiscussionController::class, 'update'])->middleware(['auth', 'admin'])->name('discussions.update');
 
-});
 
-Route::middleware('auth')->group(function () {
+
     Route::post('/discussions', DiscussionStoreController::class)->name('discussions.store');
     Route::post('/discussions/{discussion}/posts', PostStoreController::class)->name('posts.store');
     Route::delete('/discussions/{discussion}', DiscussionDestroyController::class)->name('discussions.destroy');
     Route::patch('/discussions/{discussion}/solution', DiscussionSolutionPatchController::class)->name('discussions.solution.patch');
-});
 
 
-Route::get('/discussions/{discussion:slug}', DiscussionShowController::class)->name('discussions.show');
+
+    Route::get('/discussions/{discussion:slug}', DiscussionShowController::class)->name('discussions.show');
 
 
 
@@ -115,11 +99,10 @@ Route::get('/discussions/{discussion:slug}', DiscussionShowController::class)->n
 | Post Routes
 |--------------------------------------------------------------------------
 */
-Route::prefix('posts')->middleware(['auth', 'admin'])->name('posts.')->group(function () {
-    Route::get('/', [PostController::class, 'index'])->name('index');
-    Route::get('/{post}', [PostController::class, 'show'])->name('show');
-    Route::delete('/{post}', [PostController::class, 'destroy'])->name('delete');
-});
+    Route::get('/posts', [PostController::class, 'index'])->middleware(['auth', 'admin'])->name('posts.index');
+    Route::get('/posts/{post}', [PostController::class, 'show'])->middleware(['auth', 'admin'])->name('posts.show');
+    Route::delete('/posts/delete/{post}', [PostController::class, 'destroy'])->middleware(['auth', 'admin']);
+
 
 Route::middleware('auth')->group(function () {
     Route::patch('/posts/{post}', PostPatchController::class)->name('posts.patch');
@@ -131,113 +114,16 @@ Route::middleware('auth')->group(function () {
 | Markdown Routes
 |--------------------------------------------------------------------------
 */
-Route::post('/markdown', MarkdownController::class)->name('markdown.parse');
+Route::post('/markdown', MarkdownController::class)->name('markdown');
 
 /*
 |--------------------------------------------------------------------------
 | Profile Routes
 |--------------------------------------------------------------------------
 */
-Route::prefix('profile')->name('profile.')->middleware('auth')->group(function () {
-    Route::get('/', [ProfileController::class, 'edit'])->name('edit');
-    Route::patch('/', [ProfileController::class, 'update'])->name('update');
-    Route::delete('/', [ProfileController::class, 'destroy'])->name('destroy');
-});
+    Route::get('/profile', [ProfileController::class, 'edit'])->middleware('auth')->name('profile.edit');
+    Route::patch('/profile', [ProfileController::class, 'update'])->middleware('auth')->name('profile.update');
+    Route::delete('/profile', [ProfileController::class, 'destroy'])->middleware('auth')->name('profile.destroy');
 
 
 
-
-//use App\Http\Controllers\{
-//    DiscussionDestroyController,
-//    DiscussionShowController,
-//    DiscussionSolutionPatchController,
-//    DiscussionStoreController,
-//    ForumIndexController,
-//    MarkdownController,
-//    PostDestroyController,
-//    PostController,
-//    PostPatchController,
-//    PostStoreController,
-//    ProfileController,
-//    SummaryController,
-//    UserController,
-//    TopicController,
-//    DiscussionController
-//};
-//use Illuminate\Foundation\Application;
-//use Illuminate\Support\Facades\Route;
-//use Inertia\Inertia;
-//
-///*
-//|--------------------------------------------------------------------------
-//| Web Routes
-//|--------------------------------------------------------------------------
-//|
-//| Here is where you can register web routes for your application. These
-//| routes are loaded by the RouteServiceProvider within a group which
-//| contains the "web" middleware group. Now create something great!
-//|
-//*/
-//
-//Route::get('/', function () {
-//    return Inertia::render('Welcome', [
-//        'canLogin' => Route::has('login'),
-//        'canRegister' => Route::has('register'),
-//        'laravelVersion' => Application::VERSION,
-//        'phpVersion' => PHP_VERSION,
-//    ]);
-//});
-//
-//Route::get('/', ForumIndexController::class)->name('home');
-//Route::get('/discussions/{discussion:slug}', DiscussionShowController::class)->name('discussions.show');
-//
-//// New routes for additional pages
-//Route::get('/summary', [SummaryController::class, 'index'])->name('summary');
-//
-//Route::prefix('users')->name('users.')->group(function () {
-//    Route::get('/', [UserController::class, 'index'])->name('index');
-//    Route::get('/create', [UserController::class, 'create'])->name('create');
-//    Route::post('/', [UserController::class, 'store'])->name('store');
-//    Route::get('/{user}', [UserController::class, 'show'])->name('show');
-//    Route::get('/{user}/edit', [UserController::class, 'edit'])->name('edit');
-//    Route::delete('/{user}', [UserController::class, 'destroy'])->name('destroy');
-//});
-//
-//Route::prefix('topics')->name('topics.')->group(function () {
-//    Route::get('/', [TopicController::class, 'index'])->name('index');
-//    Route::get('/create', [TopicController::class, 'create'])->name('create');
-//    Route::post('/', [TopicController::class, 'store'])->name('store');
-//    Route::get('/{topic}', [TopicController::class, 'show'])->name('show');
-//    Route::get('/{topic}/edit', [TopicController::class, 'edit'])->name('edit');
-//    Route::put('/{topic}', [TopicController::class, 'update'])->name('update');
-//    Route::delete('/{topic}', [TopicController::class, 'destroy'])->name('destroy');
-//});
-//
-//Route::prefix('discussions')->name('discussions.')->group(function () {
-//    Route::get('/', [DiscussionController::class, 'index'])->name('index');
-//    Route::get('/create', [DiscussionController::class, 'create'])->name('create');
-//    Route::post('/', [DiscussionController::class, 'store'])->name('store');
-//    Route::get('/{discussion}/edit', [DiscussionController::class, 'edit'])->name('edit');
-//    Route::put('/{discussion}', [DiscussionController::class, 'update'])->name('update');
-//    Route::delete('/{discussion}', [DiscussionController::class, 'destroy'])->name('destroy');
-//});
-//
-//Route::get('/posts', [PostController::class, 'index'])->name('posts');
-//Route::post('/markdown', MarkdownController::class)->name('markdown');
-//
-//Route::middleware('auth')->group(function () {
-//    Route::post('/discussions', DiscussionStoreController::class)->name('discussions.store');
-//    Route::post('/discussions/{discussion}/posts', PostStoreController::class)->name('posts.store');
-//    Route::delete('/discussions/{discussion}', DiscussionDestroyController::class)->name('discussions.destroy');
-//    Route::patch('/discussions/{discussion}/solution', DiscussionSolutionPatchController::class)->name('discussions.solution.patch');
-//    Route::patch('/posts/{post}', PostPatchController::class)->name('posts.patch');
-//    Route::delete('/posts/{post}', PostDestroyController::class)->name('posts.destroy');
-//
-//    Route::prefix('profile')->name('profile.')->group(function () {
-//        Route::get('/', [ProfileController::class, 'edit'])->name('edit');
-//        Route::patch('/', [ProfileController::class, 'update'])->name('update');
-//        Route::delete('/', [ProfileController::class, 'destroy'])->name('destroy');
-//    });
-//});
-//
-//require __DIR__ . '/auth.php';
